Moved changes to draw.rs
