+++ src/traits/draw.rs
@@ -1,4 +1,4 @@
-use crate::math::Point2;
+use crate::math::{Point2, Vec2};
 use crate::game::consts::PI;
 pub use graphics::{Rectangle, Context, Graphics};
 use graphics::Transformed;
@@ -329,6 +329,238 @@ impl Draw for GenericShape {
         }
 
     }
+}
+
+fn check_collision(s1: GenericShape, s2: GenericShape) -> bool {
+    
+    if let Some(s1_corners) = s1.get_corners() {
+
+        if let Some(s2_corners) = s2.get_corners() {
+            
+            let mut proj_x1: Vec<Point2> = Vec::new();
+            let mut proj_y1: Vec<Point2> = Vec::new();
+            let mut proj_x2: Vec<Point2> = Vec::new();
+            let mut proj_y2: Vec<Point2> = Vec::new();
+
+            let line = Vec2::new(s1_corners[1].x - s1_corners[0].x, s1_corners[1].y - s1_corners[0].y);
+            let norm = line.normal_unit();
+            for point in s2_corners.iter() {
+                let v = Vec2::new_from_point(s1_corners[0] - *point);
+                let py = s1_corners[0] + project(v, line);
+                let px = s1_corners[0] + project(v, norm);
+                proj_y1.push(px);
+                proj_x1.push(py);
+            }
+
+            let line = Vec2::new(s2_corners[2].x - s2_corners[0].x, s2_corners[2].y - s2_corners[0].y);
+            let norm = line.normal_unit();
+            for point in s1_corners.iter() {
+                let v = Vec2::new_from_point(s2_corners[0] - *point);
+                let py = s2_corners[0] + project(v, line);
+                let px = s2_corners[0] + project(v, norm);
+                proj_y2.push(px);
+                proj_x2.push(py);
+            }
+
+            let mut ls_x1 = [proj_x1[0], proj_x1[0]];
+            let mut ls_y1 = [proj_y1[0], proj_y1[0]];
+            let mut ls_x2 = [proj_x2[0], proj_x2[0]];
+            let mut ls_y2 = [proj_y2[0], proj_y2[0]];
+            for i in 0..4 {
+                for j in i+1..4 {
+
+                    let d = proj_x1[i] - proj_x1[j];
+                    let d = Vec2::new_from_point(d);
+
+                    let ls = ls_x1[0] - ls_x1[1];
+                    let ls = Vec2::new_from_point(ls);
+
+                    if Vec2::dot_product(d,d) > Vec2::dot_product(ls, ls) {
+                        ls_x1[0] = proj_x1[i];
+                        ls_x1[1] = proj_x1[j];
+                    }
+
+                    let d = proj_y1[i] - proj_y1[j];
+                    let d = Vec2::new_from_point(d);
+
+                    let ls = ls_y1[0] - ls_y1[1];
+                    let ls = Vec2::new_from_point(ls);
+
+                    if Vec2::dot_product(d,d) > Vec2::dot_product(ls, ls) {
+                        ls_y1[0] = proj_y1[i];
+                        ls_y1[1] = proj_y1[j];
+                    }
+
+                    let d = proj_x2[i] - proj_x2[j];
+                    let d = Vec2::new_from_point(d);
+
+                    let ls = ls_x2[0] - ls_x2[1];
+                    let ls = Vec2::new_from_point(ls);
+
+                    if Vec2::dot_product(d,d) > Vec2::dot_product(ls, ls) {
+                        ls_x2[0] = proj_x2[i];
+                        ls_x2[1] = proj_x2[j];
+                    }
+
+                    let d = proj_y2[i] - proj_y2[j];
+                    let d = Vec2::new_from_point(d);
+
+                    let ls = ls_y2[0] - ls_y2[1];
+                    let ls = Vec2::new_from_point(ls);
+
+                    if Vec2::dot_product(d,d) > Vec2::dot_product(ls, ls) {
+                        ls_y2[0] = proj_y2[i];
+                        ls_y2[1] = proj_y2[j];
+                    }
+                }
+            }
+
+            let a1 = ls_x1[0];
+            let b1 = ls_x1[1];
+
+            let a2 = ls_y1[0];
+            let b2 = ls_y1[1];
+
+            let a3 = ls_x2[0];
+            let b3 = ls_x2[1];
+
+            let a4 = ls_y2[0];
+            let b4 = ls_y2[1];
+
+            let i1 = line_intersection(s1_corners[0], s1_corners[1], a1, b1);
+            let i2 = line_intersection(s1_corners[0], s1_corners[2], a2, b2);
+            let i3 = line_intersection(s2_corners[0], s2_corners[2], a3, b3);
+            let i4 = line_intersection(s2_corners[0], s2_corners[1], a4, b4);
+            if i1 && i2 && i3 && i4 {
+                true
+            } else {
+                false
+            }
+
+        } else {
+
+            match s2.shape {
+                ShapeVariant::Circle{size: _s, radius: r} => {
+                    circle_intersection(s2.center_point(), r, s1_corners)
+                },
+                _ => false
+            }
+
+        }
+
+    } else {
+        
+        if let Some(s2_corners) = s2.get_corners() {
+
+            match s1.shape {
+                ShapeVariant::Circle{size: _s, radius: r} => {
+                    circle_intersection(s1.center_point(), r, s2_corners)
+                },
+                _ => false
+            }
+
+        } else {
+            
+            match [s1.shape, s2.shape] {
+                [ShapeVariant::Circle{size: _s1, radius: r1}, ShapeVariant::Circle{size: _s2, radius: r2}] => {
+
+                    let d = Vec2::new_from_point(s1.center_point() - s2.center_point());
+                    let d = Vec2::dot_product(d, d);
+
+                    d < (r1 + r2).powi(2) 
+
+                },
+                _ => false
+            }
+
+        }
+
+    }
+
+
+}
+
+fn project(vec: Vec2, line: Vec2) -> Point2 {
+
+    let norm = Vec2::new_unit(line.x, line.y);
+    let c = Vec2::dot_product(vec, norm) / Vec2::dot_product(norm, norm);
+
+    Point2{x: norm.x * -c, y: norm.y * -c}
+
+}
+
+fn circle_intersection(c: Point2, r: f64, corners: Vec<Point2>) -> bool {
+
+    let p1 = corners[0] + project(Vec2::new_from_point(corners[0] - c), Vec2::new(corners[1].x - corners[0].x, corners[1].y - corners[0].y));
+    let d1 = Vec2::new_from_point(p1 - c);
+    let d1 = Vec2::dot_product(d1, d1);
+
+    let p2 = corners[0] + project(Vec2::new_from_point(corners[0] - c), Vec2::new(corners[2].x - corners[0].x, corners[2].y - corners[0].y));     
+    let d2 = Vec2::new_from_point(p2 - c);
+    let d2 = Vec2::dot_product(d2, d2);
+
+    let p3 = corners[2] + project(Vec2::new_from_point(corners[0] - c), Vec2::new(corners[3].x - corners[2].x, corners[3].y - corners[2].y));
+    let d3 = Vec2::new_from_point(p3- c);
+    let d3 = Vec2::dot_product(d3, d3);
+
+    let p4 = corners[1] + project(Vec2::new_from_point(corners[0] - c), Vec2::new(corners[3].x - corners[1].x, corners[3].y - corners[1].y));
+    let d4 = Vec2::new_from_point(p4 - c);
+    let d4 = Vec2::dot_product(d4, d4);
+
+    let r_2 = r.powi(2);
+
+    // println!("{}, {}", d2 <= r_2, point_on_line(p2, corners[0], corners[2]));
+    (d1 <= r_2 && point_on_line(p1, corners[1], corners[0])) || 
+    (d2 <= r_2 && point_on_line(p2, corners[2], corners[0])) ||
+    (d3 <= r_2 && point_on_line(p3, corners[3], corners[2])) ||
+    (d4 <= r_2 && point_on_line(p4, corners[3], corners[1])) 
+
+}
+
+fn point_on_line(p: Point2, l1: Point2, l2: Point2) -> bool {
+
+    if l1.x != l2.x {
+        within(p.x, l1.x, l2.x)
+    } else {
+        within(p.y, l1.y, l2.y)
+    }
+
+}
+
+fn within(p: f64, q: f64, r: f64) -> bool{
+    (q <= p && p <= r) || (r <= p && p <= q)
+}
+
+fn line_intersection(p1: Point2, p2: Point2, q1: Point2, q2: Point2) -> bool {
+
+    let p = vec![p1,p2];
+    let q = vec![q1,q2];
+
+    let mut max_dist = 0.0;
+    for i in 0..2 {
+        for j in 0..2 {
+
+            let v = Vec2::new_from_point(p[i] - q[j]);
+            let d = Vec2::dot_product(v, v);
+            if d > max_dist {
+                max_dist = d;
+            }
+
+        }
+    }
+    
+    let dist_p = Vec2::new_from_point(p2-p1);
+    let dist_p = Vec2::dot_product(dist_p, dist_p);
+
+    let dist_q = Vec2::new_from_point(q2-q1);
+    let dist_q = Vec2::dot_product(dist_q, dist_q);
+
+    if max_dist.sqrt() <= dist_p.sqrt() + dist_q.sqrt(){
+        true
+    } else {
+        false
+    }
+
 }
